refactor(oui-datagrid): extract column builder
